fix bugs

+++ inference.py
@@ -13,7 +13,7 @@ from modules.resnet_18_baseline_fp32 import ResNet18 as ResNet18_baseline
 # --------------------------
 MODEL_PATH = "./pytorch/model/net_123.pth"
 DATA_ROOT = "./pytorch/data"
-MODEL_DTYPE = "FP32"  # 优化模型的推理精度
+MODEL_DTYPE = "FP16"  # 优化模型的推理精度
 BATCH_SIZES = [8, 16, 32, 64, 128, 256, 512]  # 需要测试的batch size列表
 NUM_WORKERS = 4
 RUN_TIMES = 1  # 每个batch size的测试轮次（仅优化模型）
